# planner: optimize the performance of `PlanCacheParamList.String()`