@hawkingrei 2023-09-04
It also fixes the #TestShowStatsLocked.